x-pack/filebeat/input/entityanalytics/provider/okta: allow fine-grain control of API requests
This adds support for specifying which of users/devices to collect from
the Okta API endpoints in order to reduce network costs for users who do
not need a full set of entities.

The current change does not change the behaviour of device collection of
registered owners and registered users; when the "devices" dataset is
selected there user entities will still be collected as they are
considered here as an attribute of the device, rather than a component
of the users dataset.