winlogbeat/eventlog: ensure event loggers retain metric collection when handling recoverable errors (#36483)
When winlogbeat's event loggers encounter recoverable errors they handle
this by closing and reopening the channel. This causes the metric
collection for the beat and dependent winlog filebeat input to lose
metric collection as metric registration only occurs on configuration.
So add a soft-close method, Reset, that only closes the channel and
leaves the event logger valid, leaving the metrics in tact, and use this
for recovering from errors.