[DOCS] Link to ZIP archive in Windows install docs (#36553)
**Problem:**
The current Windows installation docs use the Windows zip archive. The docs link to the generic download page, which also contains an msi file for Windows. If a user accidentally downloads the msi file instead, they can't complete the install using the instructions in the docs.

**Solution:**
Update the docs to link specifically to the Windows zip archive.

Closes https://github.com/elastic/beats/issues/36552