Fix race condition on TestFileWatcher (#36570)
One of the TestFileWatcher substests expects to get two FS events, one
for each file it creates, however, at least on Windows, those events
can come on any order. This commit fixes the test by looking for any
of the events and ensuring we get only the two expected events.