Fix flattened data_stream.* fields under Elastic-Agent
If the data_stream is not flattened, Beats will receive it twice, in
the Config.Source and as DataStream, that is the normal behaviour when
the Elastic-Agent is managed by Fleet and was causing a duplicated key
error.

This commit fixes that by only returning the error if the values in
Config.Source and DataStream are different.

The tests now cover this case.