Enable awscloudwatch input to collect logs from linked source accounts.
For cross account monitoring, we need to use the log group ARN, instead
of log group name in order to retreive logs from outside the monitoring
account. This change used ARN throughout (but still requires a
conversion back to get the 'plain' log group name to construct the log metadata
at the end - not ideal).