x-pack/filebeat/input/httpjson: improve template evaluation logging (#36668)
This adds valuable missing context to the debug logging emitted for
template evaluation. Previously, only the final result was printed to
the log, which failed to provide the information required to be able to
determine why any given template was being executed. So add the target
name to the logs.