packetbeat/protos{,/amqp}: fix transaction hash size constant (#36723)
The transaction hash size is currently calculated as 2⊕16 (18) which
is most likely not what was intended. The code entered the code base in
the initial import which does not have an associated review. However,
the symbol ^ signifies exponent in other languages, and XOR is almost never
what someone would use for defining a constant like this, so it is very
likely that the original author intended a 64Ki cache size.

(cherry picked from commit df10d97e8ea1d8dd948d066025076a252e688971)