x-pack/filebeat/input/entityanalytics/provider/okta: avoid work on unwanted datasets
During full sync the provider may have state from a previous dataset. So
in the case that the user has changed dataset from users to devices or
vice versa the provider may publish already existing state in the entity
graph. This change adds conditional checks to ensure that unwanted
dataset records are not published.