Optimize memory allocations in the event processing pipeline
Previously, every time a processor ran on an event we made a clone of
the entire event for two reasons:

1. this event could have some nested maps that are
shared among multiple events.

2. in case a processor fails to make a change it should be able to
revert its partial changes.

This change added a new `EventEditor` wrapper that is used for collecting
pending event changes in processors with an option to `Apply` or `Reset` them.

Additionally, this `EventEditor` takes care of the efficient memory management when
making changes to an event by cloning only the nested maps that
processors access or modify. Most of the processors just put new keys
or delete existing keys on the root-level, so most of the time the nested maps in the
event remain untouched and it does not require the whole event to be cloned.