Replace `event.batch_id` with `event.metric_names`
The `event.batch_id` with its random values is a wrong choice as a
dimension field for a time series database. It would create a new
time series at each iteration, which is terrible.

The `event.metric_names` will keep the values to a recurring set of
field names all having the same ingest delay.