[Azure Metrics] Fix CassandraConnectionClosures metric configuration (#34742)
* Move CassandraConnectionClosures metrics

The name of the metric, "CassandraConnectionClosures" is listed in the
wrong section of the metrics configuration.

It should be moved to a different section that includes metric names
which may or may not be available depending on the environment, with
the "ignore_unsupported" setting set to true.

* Update changelog

* Fix dimensions for CassandraConnectionClosures

The `CassandraConnectionClosures` metric only supports the following
dimensions:

- `ClosureReason`
- `Region`

(cherry picked from commit 42f2f9426043d5ee6fe76ec490e74c9680fdd1b0)