Auditbeat,Metricbeat - add /inputs/ to HTTP monitoring endpoint
Make metrics published by "inputs" available through the /inputs/ route
on the HTTP monitoring endpoint of Auditbeat and Metricbeat.

For Agent, include a snapshot of those metrics within the Agent diagnostics
bundle as "input_metrics.json".

When running under Agent, each module instance is configured with only a single
metricset. That module is given a unique `id`. That ID is what will be used
as the `id` within the /inputs/ data. And that `id` will also be added as context to the
logger that is passed into every metricset so that any log messages from a metricset
can be associated back to the agent stream ID).

Relates #36945