Auditbeat,Metricbeat - add /inputs/ to HTTP monitoring endpoint (#36971)
Make metrics published by "inputs" available through the /inputs/ route
on the HTTP monitoring endpoint of Auditbeat and Metricbeat.

For Agent, include a snapshot of those metrics within the Agent diagnostics
bundle as "input_metrics.json".

When running under Agent, each module instance is configured with only a single
metricset. That module is given a unique `id`. That ID is what will be used
as the `id` within the /inputs/ data. And that `id` will also be added as context to the
logger that is passed into every metricset so that any log messages from a metricset
can be associated back to the agent stream ID).

Relates #36945

Remove module and metricset keys from metricset metrics. 
For the `/inputs/` API, `input` is they key used to identify the type of "input" running.
The `module` and `metricset` keys become redundant with the addition of `input`.
I don't know of anything that relies on those fields.