[GCP] [Billing] Log an error message when the metricset finds no tables (#36775)
* Log an error message when we find no tables

Not finding any table is an error state for this metricset. It can
happen because the user made a mistake in the configuration file or
the service account lacks the needed permissions.

Regardless of the origin, it's a condition that we should report to the
user and give hints for troubleshooting.