packetbeat/{beater,protos,sniffer}: allow assignment of protocols to interfaces (#36852)
Previously the protocols association was held in a *protocols.ProtocolStruct in
common for all interfaces. This prevented individual interfaces that were only
associated with a subset of the protocols that were being captured from working
specifically on that subset. The new code moves the construction of the protocols
sets into setupSniffer where they are associated with the interface that will be
capturing them. They are then assigned to child sniffers in sniffer.New based on
the interface.

This change has no effect on the behaviour of packetbeat as it stands as configs
from packetbeat and from agent are only able to express the existence of a single
interface. Changes to configuration format/shape will enable multiple interfaces
to be enabled in packetbeat and within a single agent policy, making use of the
change here.