Switch to consumption 1.1.0 with patches
The Consumption SDK does not URL encode the query parameters in the
next link when the response requires more than one page to fetch all
the usage details items.

Our Azure SDK for Go fork contains a small patch to URL encode the
next link value and avoid the 400 error.

We also upgraded the azcore package to use the standard URL helper for
the Azure SDK.