[elastic-agent-system-metrics] add PssFreeSnapshot (#115)
## What does this PR do?

Frees the memory that is allocated when PssCaptureSnapshot is called.

## Why is it important?

memory leak

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added an entry in `CHANGELOG.md`

## Author's Checklist

<!-- Recommended
Add a checklist of things that are required to be reviewed in order to
have the PR approved
-->
- [ ] Written by hand. Need to figure out what needs to happen for "go
generate" to work. Does not work from Mac.

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the
word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- 

## Testing

- run metricbeat 8.11.0 or 8.11.1 on Windows, notice memory consumption
of metricbeat continually climbs
- rebuild metricbeat using this pr
- memory consumption rises in first 10min or so, but is steady after
that