Replace Truncate() with Round()
During a testing session on 8.11.2, we noticed some skipped
collections on one of the testing agents.

Debug information revealed the metricset skipped some collections due
to a 1-second difference between the reference time in the current
collection, and the reference time in the previous collection, making
the collection period is 1 second shorter (299s instead of 300s).

Collection skip may happen due to reference time rounding.

For example, the timestamp 2023-12-08T10:58:32.999Z may become
2023-12-08T10:58:32.000Z due to the truncation.

As of today, this problem is happening on one agent only, but the
problem is real, and we should replace the truncate(1s) with a
round(1s) to eliminate fluctuations.