Log raw events and errors containing events to a separate file
This commit introduces a new logger that can be configured through
`logging.events` that can be used to log any message that contains the
whole event or could contain the whole event.

At the moment it is used by the elasticsearch output to log indexing
errors containing the whole event and errors returned by Elasticsearch
that can potentially contain the whole event.