Lower logging level when unpacking config in autodiscover
When unpacking a config while applying the autodiscover config
template, sometimes a configuration template cannot be resolved at
that moment but will be resolved at a later stage. This commit brings
the logging level of this error back to debug as it can become too
verbose and not always means the config is invalid.