x-pack/filebeat/input/{cel,websocket}: prevent addition of regexp extension when no patterns are provided (#38181)
In the case of cel, the extension was always being added, and then was
conditionally added when a pattern was configured, resulting in a failed
configuration. In the case of websocket the conditional addition was not
being performed, so no failure occurred, but additional look-up load was
being added. So unify the approaches to the correct conditional addition.