metricbeat/module/rabbitmq/queue - Fix data type for consumers.utilisation.pct
Fix mapping type for `rabbitmq.queue.consumers.utilisation.pct` to be `scaled_float` instead of `long` because the values range on `[0, 1.0]`. Previously, truncation converted the value to an integer so it either reported zero or one.

There was a bug in the unit tests that did not catch this until we updated to testify 1.9.0 (https://github.com/stretchr/testify/pull/1531).