filebeat stateful inputs - remove source name from context.ID
For stateful inputs the context.ID is modified by appending the source name
(as returned by the input-cursor.Source#Name method). This creates an
inconsistency in behavior between stateless and stateful inputs.

I found the behavior to be surprising because the ID is user-configurable, but
the user's configured value gets mutated (and only for some inputs types). So I think
the clearest solution is to stop appending the source name.

From analyzing each of the stateful inputs that will be affected by this change,
none of them use the ID is manner that will be affected. The value only appears
to be used in log messages, metrics, or http tracer file names.

| Stateful Input                         | context.ID usages                   |
|----------------------------------------|-------------------------------------|
| filebeat/input/filestream              | log message                         |
| filebeat/input/journald                | not used                            |
| filebeat/input/winlog                  | not used                            |
| x-pack/filebeat/input/awss3            | metric ID                           |
| x-pack/filebeat/input/azureblobstorage | not used                            |
| x-pack/filebeat/input/cel              | metric ID and http tracer file name |
| x-pack/filebeat/input/gcs              | not used                            |
| x-pack/filebeat/input/httpjson         | metric ID and http tracer file name |
| x-pack/filebeat/input/o365audit        | not used                            |
| x-pack/filebeat/input/shipper          | not used                            |
| x-pack/filebeat/input/websocket        | metric ID                           |

The cursor ID that is written into the registry is constructed independent of the
context.ID so it should not be affected. See
https://github.com/elastic/beats/blob/78dc6649400610e9f908af7c90f35819f061e4e0/filebeat/input/v2/input-cursor/input.go#L171-L176