Clean up non-indexable policy handling in the ES output (#38591)
Simplify non-indexable policy handling:
- Replace the lookup table of generic factories and custom index selectors with a simple `deadLetterIndex` string, since the only valid non-indexable policies are to use a dead letter index or to drop the event if none is provided.
- Make `elasticsearch.ClientSettings` and its fields internal rather than public, and clarify some field names. (Aside from better encapsulation, this makes it easier to see that this cleanup doesn't break anything else.)

This is a preparation for the pending early-encoding change. The non-indexable policy code needs to be rewritten to support early encoding, so I split most of the cleanups out into this standalone PR to be kinder to code reviewers. This PR doesn't change any user-visible behavior.