x-pack/filebeat/module/o365: fix handling of o365.audit.OperationProperties (#38709)
In some cases, this field may be a string. This results in a mapping
failure. So in cases where the field is a string, conditionally parse
out the JSON.

Use the ingest pipeline rather than modifying the js processor due to
clarity.

(cherry picked from commit 8105935e24fb12a2120488290015a38f143d8471)

# Conflicts:
#	x-pack/filebeat/module/o365/fields.go