Dedicated script for detecting changesets on Buildkite
After the introduction of a centralized beats pipeline and statically
defined changeset rules in #38783, we still need a way to grab
changes for modules for IT tests.

Currently this is done in common.sh#defineModuleFromTheChangeSet().
This commit adds a new script that has only the changeset related
functionality, required to build the `MODULE` env var.

The functions themselves, are copied from common.sh (with minor
enhancements in documentation), paving the way to sunset common.sh
in the future.