Fix handling of truncated files for Filestream input (#38416)
This commit fixes the handling of file truncation in the Filestream input. In some file truncation cases Filebeat would detect the truncation but it would not set the offset to zero, leading to the registry offset always increasing and causing the file to be re-ingested every time Filebeat was restarted.