Fix panic when more than 32767 pipeline clients are active (#38556)
The publishing pipeline would panic when more than 32767 clients were
active, that happened because each client would add two channels to a
slice and an infinity loop would use reflect.Select on this
list. reflect.Select supports a maximum of 65536 cases, if there are
more it panics.

This PR fixes this by removing the need for this list. The pipeline
used an infinity loop to detect if the CloseRef from the pipeline
client was closed, if it happened, then it would call client.Close. By
analysing the code setting CloseRef we identified there was no need
for the pipeline to be responsible for propagating this signal, most
of the times the pipeline client was created, there was already a
defer to close it, in the few places where it was not present we added
one.

Now the pipeline is not responsible for closing any client, whomever
creates a client is responsible for correctly closing it.