Fix clean_inactive for Filestream input (#38632)
* Add tests for store GC

* Add changelog

* Run mage check and update all necessary files

* Improve changelog entry

* Fix `clean_inactive` for Filestream input

The `clean_inactive` parameter was being parsed with the wrong key. It
is parsed/used by an anonymous struct on `input-logfile/manager.go`,
there it was parsed and used as `CleanTimeout` (`clean_timeout`). This
`CleanTimeout` setting has got exactly the same effect as the
`clean_inactive` described in our documentation.

This commit fixes this bug by renaming `clean_timeout` to
`clean_inactive` so the configuration value can have effect.

* Add tests for `clean_inactive` and fix documentation

* Add changelog

* Update test config

* Add licence headers and build tags

* PR improvements

- Rename `CleanTimeout` to `CleanInactive`
- Remove commented out code

* Fix rebase conflicts

* Fix lint errors

* Jenkins test this PR