Update ECS in auditbeat sessionmd processor (#38994)
The sessionmd processor requires some of the latest process field from ECS, that are not currently in libbeat. This adds the required ECS field assets to the processor.

Without these fields, some field types would be incorrect. For example, `session_leader.start` should be date, but were sent as keyword.