Improve conditionals for x-pack/metricbeat BK CI (#39071)
PR#38913 had a bug for the extended tests group:
it wouldn't allow to run steps in the extended
group when not using a PR (e.g. merge commits).

This commit allows all steps in the extended
group to run when the trigger is not a PR.

(cherry picked from commit 8738b6cde971378e31ca910e266b40ed4c0de6cf)