Cleanup: Replace custom goroutine shutdown listeners with shared context wrapper (#38957)
Simplify context handling in the awscloudwatch and awss3 inputs.

The awscloudwatch and awss3 inputs both need to convert their Beats input context to a standard golang context for use with the AWS API, and both of them use custom wrappers that create an extra goroutine to listen for Beat shutdown. I added an explicit wrapper object to the `v2.InputContext` API that is a single function call and requires no extra goroutines, and replaced the context wrapper in both inputs.

This doesn't change any functional behavior.