Keep exit event process in DB
With the add_session_metadata processor, don't remove processes from the process
db when the process has exited.

The processor can be run on an fork/exec events after the process has actually
exited, so the process must remain in the DB after it has exited, so the info
can be used in enrichment of these events.

Now the process is kept in the DB, and the exit code is appended, so the exit
code is also now properly enriched for exit events.