Dynamic Pipeline scheduler script Buildkite
As a follow up to PR#39179 we need an additional
Buildkite pipeline that can dynamically evaluate
the release branches and trigger separate runs
per branch on a schedule.

This commit adds a Python script that generates
the dynamic steps as dictated by an env var
`PIPELINES_TO_TRIGGER`

It is intended to be used by a generic
scheduler pipeline defining schedules like:

```
schedules:
IronBank Validation:
    branch: main
    cronline: 30 02 * * *
    message: Daily trigger of ironbank-validation pipeline per branch
    env:
    PIPELINES_TO_TRIGGER: 'ironbank-validation'
`

It's similar to what was implemented (in Bash)
for Logstash in https://github.com/elastic/ingest-dev/issues/2664

We intend to leverage the container image
`docker.elastic.co/ci-agent-images/platform-ingest/buildkite-agent-beats-ci-with-hooks:0.1`
that has already prebundled (including the ruamel.yaml package) so that
there's no need for wrapping shell code.