[add_session_metadata processor] Keep exited processes in the process DB (#39173)
With the add_session_metadata processor, don't remove processes from the process db when the process has exited.

The processor can be run on an fork/exec events after the process has actually exited, so the process must remain in the DB after it has exited, so the info can be used in enrichment of these events.

Now that the process is kept in the DB, the exit code is also appended on exit events, so the exit code can be used in enrichment of the exit events.

(cherry picked from commit 964958886a0416017d3720a4a38a7805e198b7c3)