Add pipeline-scheduler pipeline in catalog.info
As a follow up to PR#39206 and PR#39171, this commit adds a new generic
scheduling pipeline in catalog-info that serves as a central point
for scheduling any other pipeline.

Unfortunately, it's not possible to specify a custom agent (k8s image)
yet at the catalog-info level[^1], therefore we still need a small
static pipeline -- empty for now -- that uploads the needed steps.

[^1]: https://github.com/elastic/ci/blob/71e83d340e3b93ab43fcf16a7a70ac33bdeec6e9/terrazzo/terrazzo/constructs/buildkite/pipelines.py#L787-L842