Fix sqs region selection (#39327)
Fix an error in region selection that was introduced in a previous cleanup (https://github.com/elastic/beats/pull/38958). When the configured region disagrees with the region detected from the queue URL, the configured region is supposed to take precedence. Due to a misreading, my code instead chose the URL region when there is a conflict.

I've broken region selection out into another helper function to make this logic easier to test, and added several unit test cases that would have caught this mistake.